Renamed ResponseSerializer to ResponseFactory
